requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
apache-beam==2.58.1
appdirs==1.4.4
array_record==0.5.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
av==12.1.0
beautifulsoup4==4.12.3
black==21.4b2
blinker==1.8.2
bpytop==1.0.68
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.8
cycler==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@e8806d607403cf0f2634d4c5ac464109fdc7d4af
dill==0.3.1.1
distro==1.9.0
dm-tree==0.1.8
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
easydict==1.13
einops==0.8.0
email_validator==2.1.2
etils==1.5.2
exceptiongroup==1.2.1
executing==2.0.1
Farama-Notifications==0.0.4
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.5
fasteners==0.19
fastjsonschema==2.20.0
ffmpy==0.3.2
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
gcsfs==2024.6.1
gdown==5.2.0
google-api-core==2.19.0
google-auth==2.30.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.49.0
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
gradio==4.36.1
gradio_client==1.0.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
gymnasium==1.0.0
h11==0.14.0
h5py==3.11.0
hdfs==2.7.3
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.26.1
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
inquirerpy==0.3.4
iopath==0.1.9
ipython==8.18.1
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.4.2
Js2Py==0.74
jsonpickle==3.2.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
lightning-utilities==0.11.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mmcv==2.2.0
mmengine==0.10.4
model-index==0.1.11
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
objsize==0.7.0
omegaconf==2.3.0
open-clip-torch==2.24.0
open3d==0.18.0
openai==1.35.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.0
opt-einsum==3.3.0
optree==0.12.1
ordered-set==4.1.0
orjson==3.10.5
oss2==2.17.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pfzy==0.3.4
pillow==10.2.0
platformdirs==4.2.2
plotly==5.22.0
portalocker==2.8.2
proglog==0.1.10
promise==2.3
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
PuLP==2.8.0
pure-eval==0.2.2
py3nvml==0.2.7
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybullet==3.2.6
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pydot==1.4.2
pydub==0.25.1
pyglet==1.5.29
Pygments==2.18.0
pyjsparser==2.7.1
pymongo==4.8.0
pyparsing==3.1.2
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.4
PyYAML==6.0.1
quadrilateral-fitter==1.8
redis==5.0.8
referencing==0.35.1
regex==2024.5.15
requests==2.28.2
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.4.2
roma==1.5.0
rpds-py==0.18.1
rsa==4.9
ruff==0.4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.2.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
supervision==0.20.0
sympy==1.12
tabulate==0.9.0
tenacity==8.4.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow-cpu==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.15.0
termcolor==2.4.0
tfds-nightly==4.9.3.dev202311230044
threadpoolctl==3.5.0
timm==1.0.7
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchmetrics==1.4.0.post0
torchvision==0.18.1+cu121
tqdm==4.65.2
traitlets==5.14.3
transformers==4.44.2
trimesh==4.4.1
triton==2.3.1
typer==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.19
uvicorn==0.30.1
uvloop==0.19.0
vertexai==1.49.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
xmltodict==0.13.0
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2
zstandard==0.23.0
